move things around, fix pages panel animation
