+++ apps/builder/app/builder/features/sidebar-left/panels/marketplace/marketplace.tsx
@@ -1,7 +1,7 @@
 import { ExtensionIcon, SpinnerIcon } from "@webstudio-is/icons";
 import { Flex, rawTheme } from "@webstudio-is/design-system";
 import type { TabContentProps } from "../../types";
-import { Header, CloseButton } from "../../header";
+import { Header, CloseButton } from "../../shared/header";
 import { Overview } from "./overview";
 import { SectionTemplates } from "./section-templates";
 import { marketplacePath } from "~/shared/router-utils";
@@ -11,7 +11,7 @@ import type { MarketplaceOverviewItem } from "~/shared/marketplace/types";
 import type { MarketplaceRouter } from "~/shared/marketplace/router";
 import { createTrpcFetchProxy } from "~/shared/remix/trpc-remix-proxy";
 import type { Project } from "@webstudio-is/project";
-import { SettingsPanel } from "../pages/settings-panel";
+import { AttachedPanel } from "../../shared/attached-panel";
 import { About } from "./about";
 
 const trpc = createTrpcFetchProxy<MarketplaceRouter>(marketplacePath);
@@ -36,49 +36,56 @@ export const TabContent = ({ onSetActiveTab }: TabContentProps) => {
   const openAboutItem = items?.find((item) => item.projectId === openAbout);
 
   return (
-    <Flex direction="column" css={{ height: "100%" }}>
-      <Header
-        title="Marketplace"
-        suffix={<CloseButton onClick={() => onSetActiveTab("none")} />}
-      />
-      {itemsLoadingState !== "idle" && (
-        <Flex justify="center" css={{ mt: "20%" }}>
-          <SpinnerIcon size={rawTheme.spacing[15]} />
-        </Flex>
-      )}
-      {activeOverviewItem && buildData ? (
-        <SectionTemplates
-          name={activeOverviewItem.name}
-          data={toWebstudioData(buildData)}
-          onOpenChange={(isOpen: boolean) => {
-            if (isOpen === false) {
-              setAciveOverviewItem(undefined);
-            }
-          }}
+    <>
+      <Flex
+        direction="column"
+        css={{
+          height: "100%",
+          position: "relative",
+          zIndex: 1,
+        }}
+      >
+        <Header
+          title="Marketplace"
+          suffix={<CloseButton onClick={() => onSetActiveTab("none")} />}
         />
-      ) : (
-        <Overview
-          items={items}
-          activeProjectId={activeOverviewItem?.projectId}
-          onSelect={(activeOverviewItem) => {
-            setAciveOverviewItem(activeOverviewItem);
-            getBuildData({ projectId: activeOverviewItem.projectId });
-          }}
-          openAbout={openAbout}
-          onOpenAbout={setopenAbout}
-        />
-      )}
-      {openAboutItem && (
-        <SettingsPanel isOpen>
-          <About
-            item={openAboutItem}
-            onClose={() => {
-              setopenAbout(undefined);
+        {itemsLoadingState !== "idle" && (
+          <Flex justify="center" css={{ mt: "20%" }}>
+            <SpinnerIcon size={rawTheme.spacing[15]} />
+          </Flex>
+        )}
+        {activeOverviewItem && buildData ? (
+          <SectionTemplates
+            name={activeOverviewItem.name}
+            data={toWebstudioData(buildData)}
+            onOpenChange={(isOpen: boolean) => {
+              if (isOpen === false) {
+                setAciveOverviewItem(undefined);
+              }
             }}
           />
-        </SettingsPanel>
-      )}
-    </Flex>
+        ) : (
+          <Overview
+            items={items}
+            activeProjectId={activeOverviewItem?.projectId}
+            onSelect={(activeOverviewItem) => {
+              setAciveOverviewItem(activeOverviewItem);
+              getBuildData({ projectId: activeOverviewItem.projectId });
+            }}
+            openAbout={openAbout}
+            onOpenAbout={setopenAbout}
+          />
+        )}
+      </Flex>
+      <AttachedPanel isOpen={openAboutItem !== undefined}>
+        <About
+          item={openAboutItem}
+          onClose={() => {
+            setopenAbout(undefined);
+          }}
+        />
+      </AttachedPanel>
+    </>
   );
 };
